MVCを考慮した設計に修正
