+++ src/main/java/raisetech/studentManagement/StudentManagementApplication.java
@@ -6,6 +6,9 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
+import raisetech.studentManagement.data.Student;
+import raisetech.studentManagement.data.StudentCourse;
+import raisetech.studentManagement.repository.StudentRepository;
 
 @SpringBootApplication
 @RestController
@@ -18,32 +21,4 @@ public class StudentManagementApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(StudentManagementApplication.class, args);
 	}
-
-	@GetMapping("/studentList")
-	public List<Student> getStudentInfo() {
-		return repository.search();
-	}
-	@GetMapping("/coursesList")
-	public List<StudentCourse> getStudentCouses(){
-		return repository.searchCourses();
-	}
-/* 講義２３　では　使わないから1回消してるため　コメントアウト
-	@PostMapping("/student")
-	public void registerStudent(String name, int age) {
-		repository.registerStudent(name, age);
-	}
-
-	@PatchMapping("/student")
-	public void updateStudentName(String name, int age) {
-		repository.updateStudent(name, age);
-	}
-
-	@DeleteMapping("/student")
-	public void deleteStudent(String name) {
-		repository.deleteStudent(name);
-
-	}
-
- */
-
 }
