+++ src/main/java/raisetech/studentManagement/data/StudentCourse.java
@@ -1,4 +1,4 @@
-package raisetech.studentManagement;
+package raisetech.studentManagement.data;
 
 import java.util.Date;
 import lombok.Getter;
